Fix linux warnings Review URL: http://codereview.appspot.com/6354093/
git-svn-id: http://skia.googlecode.com/svn/trunk@4554 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ src/gpu/gl/GrGLPath.cpp
@@ -29,7 +29,7 @@ inline GrGLubyte verb_to_gl_path_cmd(const SkPath::Verb verb) {
     GR_STATIC_ASSERT(3 == SkPath::kCubic_Verb);
     GR_STATIC_ASSERT(4 == SkPath::kClose_Verb);
 
-    GrAssert(verb >= 0 && verb < GR_ARRAY_COUNT(gTable));
+    GrAssert(verb >= 0 && (size_t)verb < GR_ARRAY_COUNT(gTable));
     return gTable[verb];
 }
 
@@ -47,7 +47,7 @@ inline int num_pts(const SkPath::Verb verb) {
     GR_STATIC_ASSERT(3 == SkPath::kCubic_Verb);
     GR_STATIC_ASSERT(4 == SkPath::kClose_Verb);
 
-    GrAssert(verb >= 0 && verb < GR_ARRAY_COUNT(gTable));
+    GrAssert(verb >= 0 && (size_t)verb < GR_ARRAY_COUNT(gTable));
     return gTable[verb];
 }
 }
